Fix package of Role class. Move from org.n52.sos to org.n52.sos.ogc.sensorML.
